pyproject.toml
[project]
name = "workflow-scratch-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.0",
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "openai>=2.0.1",
    "pydantic-settings>=2.11.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.32.5",
    "sqlalchemy[postgresql-asyncpg]>=2.0.43",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
    "pydeps>=3.0.1",
    "pre-commit>=4.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py311"
fix = true
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true

